scripts: add zp_utils shared helper module; refactor 14 build scripts
