+++ scripts/build_zpb.py
@@ -1,182 +1,9 @@
 """
 Build ZP-B: p-Adic Topology (v1.4)
 """
-import os, sys
-sys.stdout.reconfigure(encoding='utf-8')
-from reportlab.lib import colors
-from reportlab.lib.pagesizes import LETTER
-from reportlab.lib.units import inch
-from reportlab.lib.styles import ParagraphStyle
-from reportlab.platypus import (SimpleDocTemplate, Paragraph, Spacer,
-                                 Table, TableStyle, PageBreak, HRFlowable)
-from reportlab.pdfbase import pdfmetrics
-from reportlab.pdfbase.ttfonts import TTFont
 
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-FONT_DIR = os.path.join(PROJECT_ROOT, '.claude-local', 'fonts') + os.sep
-pdfmetrics.registerFont(TTFont('DV',    FONT_DIR + 'DejaVuSans.ttf'))
-pdfmetrics.registerFont(TTFont('DV-B',  FONT_DIR + 'DejaVuSans-Bold.ttf'))
-pdfmetrics.registerFont(TTFont('DV-I',  FONT_DIR + 'DejaVuSans-Oblique.ttf'))
-pdfmetrics.registerFont(TTFont('DV-BI', FONT_DIR + 'DejaVuSans-BoldOblique.ttf'))
-pdfmetrics.registerFont(TTFont('DVS',   FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-B', FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-I', FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-BI',FONT_DIR + 'STIXTwo-Math.ttf'))
-
-BLUE      = colors.HexColor('#2E75B6')
-BLUE_LITE = colors.HexColor('#D5E8F0')
-GREY_LITE = colors.HexColor('#F5F5F5')
-BLACK     = colors.black
-WHITE     = colors.white
-COMP_BLUE  = colors.HexColor('#4D89C0')
-COMP_GREEN = colors.HexColor('#4D9050')
-COMP_SLATE = colors.HexColor('#60727B')
-COMP_AMBER = colors.HexColor('#BB8C26')
-SLATE_LITE = colors.HexColor('#ECEFF1')
-AMBER_LITE = colors.HexColor('#FFF8E7')
-
-TW = 6.5 * inch
-LM = RM = 1.0 * inch
-TM = BM = 1.0 * inch
-
-S = {
-    'title':    ParagraphStyle('title',    fontName='DV-B',  fontSize=18, leading=24,
-                               spaceAfter=6,  alignment=1),
-    'subtitle': ParagraphStyle('subtitle', fontName='DV-I',  fontSize=11, leading=15,
-                               spaceAfter=4,  alignment=1),
-    'h1':       ParagraphStyle('h1',       fontName='DV-B',  fontSize=13, leading=18,
-                               spaceBefore=14, spaceAfter=5, textColor=BLUE),
-    'h2':       ParagraphStyle('h2',       fontName='DV-B',  fontSize=11, leading=15,
-                               spaceBefore=10, spaceAfter=4, textColor=BLUE),
-    'h3':       ParagraphStyle('h3',       fontName='DV-B',  fontSize=10, leading=14,
-                               spaceBefore=8,  spaceAfter=3),
-    'body':     ParagraphStyle('body',     fontName='DVS',   fontSize=10, leading=14,
-                               spaceAfter=6),
-    'bodyI':    ParagraphStyle('bodyI',    fontName='DVS-I', fontSize=10, leading=14,
-                               spaceAfter=6),
-    'label':    ParagraphStyle('label',    fontName='DV-B',  fontSize=9,  leading=13,
-                               textColor=WHITE),
-    'cell':     ParagraphStyle('cell',     fontName='DVS',   fontSize=9,  leading=13),
-    'cellB':    ParagraphStyle('cellB',    fontName='DVS-B', fontSize=9,  leading=13),
-    'cellSans': ParagraphStyle('cellSans', fontName='DV',    fontSize=9,  leading=13),
-    'cellI':    ParagraphStyle('cellI',    fontName='DVS-I', fontSize=9,  leading=13),
-    'footer':   ParagraphStyle('footer',   fontName='DV-I',  fontSize=8,  leading=10,
-                               textColor=colors.grey, alignment=1),
-    'note':     ParagraphStyle('note',     fontName='DVS-I', fontSize=9,  leading=13,
-                               spaceAfter=4),
-}
-
-def sp(n=6):
-    return Spacer(1, n)
-
-def chk():
-    return '<font name="DV">&#10003;</font>'
-
-def fix(text):
-    sub_map = {'₀':'0','₁':'1','₂':'2','₃':'3','₄':'4',
-               '₅':'5','₆':'6','₇':'7','₈':'8','₉':'9',
-               'ₙ':'n','ₖ':'k','ₘ':'m','ᵢ':'i','ⱼ':'j',
-               '₊':'+','₋':'-'}
-    sup_map = {'⁰':'0','¹':'1','²':'2','³':'3','⁴':'4',
-               '⁵':'5','⁶':'6','⁷':'7','⁸':'8','⁹':'9',
-               'ⁿ':'n','ᵏ':'k','∞':'&#8734;'}
-    for ch, rep in sub_map.items():
-        text = text.replace(ch, f'<sub>{rep}</sub>')
-    for ch, rep in sup_map.items():
-        text = text.replace(ch, f'<super>{rep}</super>')
-    text = text.replace('✓', '<font name="DV">&#10003;</font>')
-    text = text.replace('✗', '<font name="DV">&#10007;</font>')
-    text = text.replace('∅', '<font name="DV">&#8709;</font>')
-    replacements = [
-        ('⊥', '&#8869;'), ('∨', '&#8744;'), ('∧', '&#8743;'),
-        ('≤', '&#8804;'), ('≥', '&#8805;'), ('≠', '&#8800;'),
-        ('∈', '&#8712;'), ('∉', '&#8713;'), ('⊆', '&#8838;'),
-        ('∪', '&#8746;'), ('∩', '&#8745;'), ('∀', '&#8704;'),
-        ('∃', '&#8707;'), ('∞', '&#8734;'), ('∑', '&#8721;'),
-        ('→', '&#8594;'), ('←', '&#8592;'), ('↔', '&#8596;'),
-        ('⇒', '&#8658;'), ('⟺', '&#10234;'), ('⟹', '&#10233;'),
-        ('⟨', '&#10216;'), ('⟩', '&#10217;'), ('‖', '&#8214;'),
-        ('—', '&#8212;'), ('–', '&#8211;'), ('·', '&#183;'),
-        ('×', '&#215;'), ('−', '&#8722;'), ('≡', '&#8801;'),
-        ('∘', '&#8728;'), ('⊗', '&#8855;'),
-        ('ε', '&#949;'), ('α', '&#945;'), ('β', '&#946;'),
-        ('γ', '&#947;'), ('δ', '&#948;'), ('η', '&#951;'),
-        ('σ', '&#963;'), ('π', '&#960;'),
-        ('Δ', '&#916;'), ('Σ', '&#931;'), ('Γ', '&#915;'),
-        ('ℚ', '&#8474;'), ('ℤ', '&#8484;'), ('ℂ', '&#8450;'),
-        ('ℕ', '&#8469;'), ('ℝ', '&#8477;'),
-        ('∥', '&#8741;'), ('⊕', '&#8853;'),
-    ]
-    for char, entity in replacements:
-        if char not in text:
-            continue
-        text = text.replace(char, entity)
-    return text
-
-def body(text, style='body'):
-    return Paragraph(fix(text), S[style])
-
-def label_box(title, rows_list):
-    data = [[Paragraph(fix(title), S['label'])]]
-    for r in rows_list:
-        data.append([Paragraph(fix(r), S['cell'])])
-    ts = TableStyle([
-        ('BACKGROUND',   (0,0), (-1,0),  BLUE),
-        ('BACKGROUND',   (0,1), (-1,-1), GREY_LITE),
-        ('TEXTCOLOR',    (0,0), (-1,0),  WHITE),
-        ('BOX',          (0,0), (-1,-1), 0.5, BLUE),
-        ('LINEBELOW',    (0,0), (-1,0),  0.5, BLUE),
-        ('LINEBEFORE',   (0,0), (-1,-1), 0.5, colors.HexColor('#CCCCCC')),
-        ('LINEAFTER',    (0,0), (-1,-1), 0.5, colors.HexColor('#CCCCCC')),
-        ('LINEBELOW',    (0,1), (-1,-2), 0.5, colors.HexColor('#CCCCCC')),
-        ('TOPPADDING',   (0,0), (-1,-1), 5),
-        ('BOTTOMPADDING',(0,0), (-1,-1), 5),
-        ('LEFTPADDING',  (0,0), (-1,-1), 8),
-        ('RIGHTPADDING', (0,0), (-1,-1), 8),
-        ('VALIGN',       (0,0), (-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=[TW], repeatRows=1)
-    t.setStyle(ts)
-    return t
-
-def data_table(headers, rows_data, col_widths, header_bg=BLUE):
-    hdr_row = [Paragraph(fix(h), S['label']) for h in headers]
-    data = [hdr_row]
-    for row in rows_data:
-        data.append([Paragraph(fix(str(c)), S['cell']) for c in row])
-    ts = TableStyle([
-        ('BACKGROUND',    (0,0), (-1,0),  header_bg),
-        ('BACKGROUND',    (0,1), (-1,-1), GREY_LITE),
-        ('ROWBACKGROUNDS',(0,1), (-1,-1), [WHITE, GREY_LITE]),
-        ('BOX',           (0,0), (-1,-1), 0.5, BLUE),
-        ('LINEBELOW',     (0,0), (-1,0),  0.5, BLUE),
-        ('INNERGRID',     (0,1), (-1,-1), 0.3, colors.HexColor('#CCCCCC')),
-        ('LINEAFTER',     (0,0), (-1,-1), 0.3, colors.HexColor('#CCCCCC')),
-        ('TOPPADDING',    (0,0), (-1,-1), 4),
-        ('BOTTOMPADDING', (0,0), (-1,-1), 4),
-        ('LEFTPADDING',   (0,0), (-1,-1), 6),
-        ('RIGHTPADDING',  (0,0), (-1,-1), 6),
-        ('VALIGN',        (0,0), (-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=col_widths, repeatRows=1)
-    t.setStyle(ts)
-    return t
-
-def make_doc(path, title_str, doc_id, version_str):
-    def footer_cb(canvas, doc):
-        canvas.saveState()
-        canvas.setFont('DV-I', 8)
-        canvas.setFillColor(colors.grey)
-        footer_text = f"Zero Paradox {doc_id}  |  {version_str}  |  April 2026  |  Page {doc.page}"
-        canvas.drawCentredString(LETTER[0]/2, 0.6*inch, footer_text)
-        canvas.restoreState()
-    return SimpleDocTemplate(
-        path, pagesize=LETTER,
-        leftMargin=LM, rightMargin=RM, topMargin=TM, bottomMargin=BM,
-        title=title_str, author="Zero Paradox Project",
-        onFirstPage=footer_cb, onLaterPages=footer_cb,
-    )
-
+import os
+from zp_utils import *
 
 def build():
     out_path = os.path.join(PROJECT_ROOT, 'ZP-B_pAdic_Topology_v1_4.pdf')
