+++ scripts/build_zpi_companion.py
@@ -10,157 +10,9 @@ New in v1.0: t_iz_r1_t3_geometric_bound (geometric norm bound from R1+T3).
 """
 
 import os
-import math
-from reportlab.lib import colors
-from reportlab.lib.pagesizes import LETTER
-from reportlab.lib.units import inch
-from reportlab.lib.styles import ParagraphStyle
-from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
-from reportlab.pdfbase import pdfmetrics
-from reportlab.pdfbase.ttfonts import TTFont
+from zp_utils import *
 from reportlab.graphics.shapes import Drawing, Line, String, Rect, Circle, Polygon
 from reportlab.graphics import renderPDF
-
-FONT_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'fonts') + os.sep
-pdfmetrics.registerFont(TTFont('DV',     FONT_DIR + 'DejaVuSans.ttf'))
-pdfmetrics.registerFont(TTFont('DV-B',   FONT_DIR + 'DejaVuSans-Bold.ttf'))
-pdfmetrics.registerFont(TTFont('DV-I',   FONT_DIR + 'DejaVuSans-Oblique.ttf'))
-pdfmetrics.registerFont(TTFont('DV-BI',  FONT_DIR + 'DejaVuSans-BoldOblique.ttf'))
-pdfmetrics.registerFont(TTFont('DVS',    FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-B',  FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-I',  FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-BI', FONT_DIR + 'STIXTwo-Math.ttf'))
-
-# Companion palette
-COMP_BLUE  = colors.HexColor('#4D89C0')
-COMP_GREEN = colors.HexColor('#4D9050')
-COMP_SLATE = colors.HexColor('#60727B')
-COMP_AMBER = colors.HexColor('#BB8C26')
-SLATE_LITE = colors.HexColor('#ECEFF1')
-AMBER_LITE = colors.HexColor('#FFF8E7')
-INDIGO     = colors.HexColor('#3949AB')
-INDIGO_LITE= colors.HexColor('#E8EAF6')
-GREEN_DARK = colors.HexColor('#1B5E20')
-GREEN_LITE = colors.HexColor('#E8F5E9')
-BLACK      = colors.black
-WHITE      = colors.white
-
-TW = 6.5 * inch
-LM = RM = 1.0 * inch
-TM = BM = 1.0 * inch
-
-CS = {
-    'title':    ParagraphStyle('ctitle',   fontName='DV-B',  fontSize=22, leading=28,
-                               alignment=1, spaceAfter=4, textColor=BLACK),
-    'subtitle': ParagraphStyle('csubtitle',fontName='DV-I',  fontSize=12, leading=16,
-                               alignment=1, spaceAfter=4),
-    'meta':     ParagraphStyle('cmeta',    fontName='DV',    fontSize=9,  leading=13,
-                               alignment=1, spaceAfter=8, textColor=colors.grey),
-    'disc':     ParagraphStyle('cdisc',    fontName='DVS-I', fontSize=9,  leading=13,
-                               spaceAfter=10, textColor=colors.HexColor('#555555')),
-    'h1':       ParagraphStyle('ch1',      fontName='DV-B',  fontSize=13, leading=17,
-                               spaceBefore=14, spaceAfter=5, textColor=COMP_BLUE),
-    'body':     ParagraphStyle('cbody',    fontName='DVS',   fontSize=10, leading=14,
-                               spaceAfter=6),
-    'caption':  ParagraphStyle('ccaption', fontName='DVS-I', fontSize=9,  leading=12,
-                               spaceAfter=8, textColor=colors.HexColor('#555555')),
-    'ex_title': ParagraphStyle('cex_title',fontName='DV-B',  fontSize=9,  leading=13,
-                               textColor=COMP_AMBER),
-    'ex_body':  ParagraphStyle('cex_body', fontName='DVS',   fontSize=9,  leading=13),
-    'rem':      ParagraphStyle('crem',     fontName='DVS-I', fontSize=9,  leading=13),
-    'kr_hdr':   ParagraphStyle('ckr_hdr',  fontName='DVS-B', fontSize=9,  leading=13,
-                               textColor=WHITE),
-    'kr_body':  ParagraphStyle('ckr_body', fontName='DVS',   fontSize=9,  leading=13),
-    'tbl_hdr':  ParagraphStyle('ctbl_hdr', fontName='DV-B',  fontSize=9,  leading=13,
-                               textColor=WHITE),
-    'tbl_body': ParagraphStyle('ctbl_body',fontName='DVS',   fontSize=9,  leading=13),
-}
-
-def sp(n=6):
-    return Spacer(1, n)
-
-def fix(text):
-    sub_map = {'₀': '0', '₁': '1', '₂': '2', '₃': '3',
-               '₄': '4', '₅': '5', '₆': '6', '₇': '7',
-               '₈': '8', '₉': '9',
-               'ₙ': 'n', 'ₖ': 'k', 'ₘ': 'm'}
-    for ch, rep in sub_map.items():
-        text = text.replace(ch, f'<sub>{rep}</sub>')
-    text = text.replace('✓', '<font name="DV">&#10003;</font>')
-    text = text.replace('∅', '<font name="DV">&#8709;</font>')
-    for char, entity in [('ℚ', '&#8474;'), ('ℤ', '&#8484;'),
-                          ('ℂ', '&#8450;'), ('ℕ', '&#8469;'),
-                          ('ℝ', '&#8477;')]:
-        text = text.replace(char, f'<font name="DV">{entity}</font>')
-    replacements = [
-        ('⊥', '&#8869;'), ('∨', '&#8744;'), ('≤', '&#8804;'),
-        ('≥', '&#8805;'), ('≠', '&#8800;'), ('∈', '&#8712;'),
-        ('→', '&#8594;'), ('←', '&#8592;'), ('∞', '&#8734;'),
-        ('⟨', '&#10216;'), ('⟩', '&#10217;'), ('—', '&#8212;'),
-        ('–', '&#8211;'), ('×', '&#215;'), ('−', '&#8722;'),
-        ('ε', '&#949;'), ('ι', '&#953;'), ('∀', '&#8704;'),
-        ('∃', '&#8707;'), ('∘', '&#8728;'), ('∥', '&#8741;'),
-        ('‖', '&#8214;'), ('α', '&#945;'), ('ω', '&#969;'),
-        ('Ω', '&#937;'), ('⇒', '&#8658;'), ('′', '&#8242;'),
-        ('⋅', '&#8901;'),
-    ]
-    for char, entity in replacements:
-        text = text.replace(char, entity)
-    return text
-
-def cbody(text):
-    return Paragraph(fix(text), CS['body'])
-
-def ccaption(text):
-    return Paragraph(fix(text), CS['caption'])
-
-def example_box(title, rows):
-    data = [[Paragraph(title, CS['ex_title'])]]
-    for r in rows:
-        data.append([Paragraph(fix(r), CS['ex_body'])])
-    ts = TableStyle([
-        ('BOX',           (0,0), (-1,-1), 1.5, COMP_AMBER),
-        ('LINEBELOW',     (0,0), (-1,0),  0.5, COMP_AMBER),
-        ('BACKGROUND',    (0,0), (-1,-1), AMBER_LITE),
-        ('TOPPADDING',    (0,0), (-1,-1), 5),
-        ('BOTTOMPADDING', (0,0), (-1,-1), 5),
-        ('LEFTPADDING',   (0,0), (-1,-1), 8),
-        ('RIGHTPADDING',  (0,0), (-1,-1), 8),
-        ('VALIGN',        (0,0), (-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=[TW])
-    t.setStyle(ts)
-    return t
-
-def remember_box(text):
-    ts = TableStyle([
-        ('BACKGROUND',    (0,0), (-1,-1), SLATE_LITE),
-        ('BOX',           (0,0), (-1,-1), 0.5, COMP_SLATE),
-        ('TOPPADDING',    (0,0), (-1,-1), 8),
-        ('BOTTOMPADDING', (0,0), (-1,-1), 8),
-        ('LEFTPADDING',   (0,0), (-1,-1), 10),
-        ('RIGHTPADDING',  (0,0), (-1,-1), 10),
-    ])
-    t = Table([[Paragraph(fix(text), CS['rem'])]], colWidths=[TW])
-    t.setStyle(ts)
-    return t
-
-def key_result_box(title, body_text):
-    data = [[Paragraph(fix(title), CS['kr_hdr'])],
-            [Paragraph(fix(body_text), CS['kr_body'])]]
-    ts = TableStyle([
-        ('BACKGROUND',    (0,0), (-1,0),  COMP_GREEN),
-        ('BACKGROUND',    (0,1), (-1,-1), colors.white),
-        ('BOX',           (0,0), (-1,-1), 0.5, COMP_GREEN),
-        ('TOPPADDING',    (0,0), (-1,-1), 6),
-        ('BOTTOMPADDING', (0,0), (-1,-1), 6),
-        ('LEFTPADDING',   (0,0), (-1,-1), 8),
-        ('RIGHTPADDING',  (0,0), (-1,-1), 8),
-        ('VALIGN',        (0,0), (-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=[TW])
-    t.setStyle(ts)
-    return t
 
 def lean_status_box(rows):
     data = [[Paragraph('Lean 4 Verification Status (ZPI.lean v1.1 — all proofs filled)',
@@ -418,7 +270,7 @@ def cycle_diagram():
 
 
 def build():
-    out_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
+    out_path = os.path.join(PROJECT_ROOT,
                             'ZP-I_Illustrated_Companion.pdf')
 
     def footer_cb(canvas, doc):
