+++ scripts/build_zpj_companion.py
@@ -7,124 +7,9 @@ Formal doc: ZP-J Self-Reference v1.0.
 """
 
 import os
-from reportlab.lib import colors
-from reportlab.lib.pagesizes import LETTER
-from reportlab.lib.units import inch
-from reportlab.lib.styles import ParagraphStyle
-from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
-from reportlab.pdfbase import pdfmetrics
-from reportlab.pdfbase.ttfonts import TTFont
+from zp_utils import *
 from reportlab.graphics.shapes import Drawing, Line, String, Rect, Circle
 from reportlab.graphics import renderPDF
-
-FONT_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'fonts') + os.sep
-pdfmetrics.registerFont(TTFont('DV',     FONT_DIR + 'DejaVuSans.ttf'))
-pdfmetrics.registerFont(TTFont('DV-B',   FONT_DIR + 'DejaVuSans-Bold.ttf'))
-pdfmetrics.registerFont(TTFont('DV-I',   FONT_DIR + 'DejaVuSans-Oblique.ttf'))
-pdfmetrics.registerFont(TTFont('DV-BI',  FONT_DIR + 'DejaVuSans-BoldOblique.ttf'))
-pdfmetrics.registerFont(TTFont('DVS',    FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-B',  FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-I',  FONT_DIR + 'STIXTwo-Math.ttf'))
-pdfmetrics.registerFont(TTFont('DVS-BI', FONT_DIR + 'STIXTwo-Math.ttf'))
-
-COMP_BLUE  = colors.HexColor('#4D89C0')
-COMP_GREEN = colors.HexColor('#4D9050')
-COMP_SLATE = colors.HexColor('#60727B')
-COMP_AMBER = colors.HexColor('#BB8C26')
-SLATE_LITE = colors.HexColor('#ECEFF1')
-AMBER_LITE = colors.HexColor('#FFF8E7')
-INDIGO     = colors.HexColor('#3949AB')
-INDIGO_LITE= colors.HexColor('#E8EAF6')
-BLACK      = colors.black
-WHITE      = colors.white
-GREY_TEXT  = colors.HexColor('#555555')
-
-TW = 6.5 * inch
-LM = RM = TM = BM = 1.0 * inch
-
-CS = {
-    'title':    ParagraphStyle('ctitle',   fontName='DV-B',  fontSize=22, leading=28,
-                               alignment=1, spaceAfter=4, textColor=BLACK),
-    'subtitle': ParagraphStyle('csubtitle',fontName='DV-I',  fontSize=12, leading=16,
-                               alignment=1, spaceAfter=4),
-    'meta':     ParagraphStyle('cmeta',    fontName='DV',    fontSize=9,  leading=13,
-                               alignment=1, spaceAfter=8, textColor=colors.grey),
-    'disc':     ParagraphStyle('cdisc',    fontName='DVS-I', fontSize=9,  leading=13,
-                               spaceAfter=10, textColor=GREY_TEXT),
-    'h1':       ParagraphStyle('ch1',      fontName='DV-B',  fontSize=13, leading=17,
-                               spaceBefore=14, spaceAfter=5, textColor=COMP_BLUE),
-    'body':     ParagraphStyle('cbody',    fontName='DVS',   fontSize=10, leading=14,
-                               spaceAfter=6),
-    'caption':  ParagraphStyle('ccaption', fontName='DVS-I', fontSize=9,  leading=12,
-                               spaceAfter=8, textColor=GREY_TEXT),
-    'ex_title': ParagraphStyle('cex_title',fontName='DV-B',  fontSize=9,  leading=13,
-                               textColor=COMP_AMBER),
-    'ex_body':  ParagraphStyle('cex_body', fontName='DVS',   fontSize=9,  leading=13),
-    'rem':      ParagraphStyle('crem',     fontName='DVS-I', fontSize=9,  leading=13),
-    'kr_hdr':   ParagraphStyle('ckr_hdr',  fontName='DVS-B', fontSize=9,  leading=13,
-                               textColor=WHITE),
-    'kr_body':  ParagraphStyle('ckr_body', fontName='DVS',   fontSize=9,  leading=13),
-    'tbl_hdr':  ParagraphStyle('ctbl_hdr', fontName='DV-B',  fontSize=9,  leading=13,
-                               textColor=WHITE),
-    'tbl_cell': ParagraphStyle('ctbl_cell',fontName='DVS',   fontSize=9,  leading=13),
-}
-
-def sp(n=6): return Spacer(1, n)
-
-def fix(text):
-    sub_map = {'₀':'0','₁':'1','₂':'2','₃':'3','₄':'4',
-               '₅':'5','₆':'6','₇':'7','₈':'8','₉':'9','ₙ':'n'}
-    for ch, rep in sub_map.items():
-        text = text.replace(ch, f'<sub>{rep}</sub>')
-    for char, ent in [('⊥','&#8869;'),('∨','&#8744;'),('∧','&#8743;'),
-                      ('≤','&#8804;'),('≥','&#8805;'),('≠','&#8800;'),
-                      ('∈','&#8712;'),('∉','&#8713;'),('⊆','&#8838;'),
-                      ('∀','&#8704;'),('∃','&#8707;'),('→','&#8594;'),
-                      ('↔','&#8596;'),('∞','&#8734;'),('—','&#8212;'),
-                      ('ε','&#949;'),('ω','&#969;'),('≡','&#8801;')]:
-        text = text.replace(char, ent)
-    text = text.replace('✓', '<font name="DV">&#10003;</font>')
-    return text
-
-def cbody(t): return Paragraph(fix(t), CS['body'])
-def ccaption(t): return Paragraph(fix(t), CS['caption'])
-
-def example_box(title, rows):
-    data = [[Paragraph(title, CS['ex_title'])]]
-    for r in rows:
-        data.append([Paragraph(fix(r), CS['ex_body'])])
-    ts = TableStyle([
-        ('BOX',          (0,0),(-1,-1), 1.5, COMP_AMBER),
-        ('LINEBELOW',    (0,0),(-1,0),  0.5, COMP_AMBER),
-        ('BACKGROUND',   (0,0),(-1,-1), AMBER_LITE),
-        ('TOPPADDING',   (0,0),(-1,-1), 5), ('BOTTOMPADDING',(0,0),(-1,-1), 5),
-        ('LEFTPADDING',  (0,0),(-1,-1), 8), ('RIGHTPADDING', (0,0),(-1,-1), 8),
-        ('VALIGN',       (0,0),(-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=[TW]); t.setStyle(ts); return t
-
-def remember_box(text):
-    ts = TableStyle([
-        ('BACKGROUND',   (0,0),(-1,-1), SLATE_LITE),
-        ('BOX',          (0,0),(-1,-1), 0.5, COMP_SLATE),
-        ('TOPPADDING',   (0,0),(-1,-1), 8), ('BOTTOMPADDING',(0,0),(-1,-1), 8),
-        ('LEFTPADDING',  (0,0),(-1,-1), 10),('RIGHTPADDING', (0,0),(-1,-1), 10),
-    ])
-    t = Table([[Paragraph(fix(text), CS['rem'])]], colWidths=[TW])
-    t.setStyle(ts); return t
-
-def key_result_box(title, body_text):
-    data = [[Paragraph(fix(title), CS['kr_hdr'])],
-            [Paragraph(fix(body_text), CS['kr_body'])]]
-    ts = TableStyle([
-        ('BACKGROUND',   (0,0),(-1,0),  COMP_GREEN),
-        ('BACKGROUND',   (0,1),(-1,-1), WHITE),
-        ('BOX',          (0,0),(-1,-1), 0.5, COMP_GREEN),
-        ('TOPPADDING',   (0,0),(-1,-1), 6), ('BOTTOMPADDING',(0,0),(-1,-1), 6),
-        ('LEFTPADDING',  (0,0),(-1,-1), 8), ('RIGHTPADDING', (0,0),(-1,-1), 8),
-        ('VALIGN',       (0,0),(-1,-1), 'TOP'),
-    ])
-    t = Table(data, colWidths=[TW]); t.setStyle(ts); return t
 
 def quine_atom_diagram():
     """Diagram showing ⊥ ∈ ⊥ — the self-containing null state."""
@@ -201,7 +86,7 @@ def three_way_table():
     t.setStyle(ts); return t
 
 def build():
-    out_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
+    out_path = os.path.join(PROJECT_ROOT,
                             'ZP-J_Illustrated_Companion.pdf')
 
     def footer_cb(canvas, doc):
